requirements.txt
playwright
pydantic
python-dotenv
google-generativeai
